Timer Function Added to SEDA
The make_timer() function in this file, later renamed count_nums(), records how long SEDA (Spam Email Detective Agency) runs. It prints a statement at the very end with the result. Pretty interesting! Include this function in your code and see what the outcome is. The lines of code you should copy and paste are indicated by numbers (1) (2) and (3), and should be referenced in that order :)